# internal/apidocs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Administrative REST API for managing the MCP Data Platform.\nEndpoints cover system info, configuration, personas, auth keys, audit logs, and knowledge management.",
        "title": "MCP Data Platform Admin API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1/admin",
    "paths": {
...
        "/connection-instances": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns all database-managed connection instances.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Connections"
                ],
                "summary": "List connection instances",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/platform.ConnectionInstance"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/admin.problemDetail"
                        }
                    }
                }
            }
        },
        "/connection-instances/{kind}/{name}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns a single database-managed connection instance by kind and name.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Connections"
                ],
                "summary": "Get connection instance",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Toolkit kind (trino, datahub, s3)",
                        "name": "kind",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Instance name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/platform.ConnectionInstance"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/admin.problemDetail"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/admin.problemDetail"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Creates or updates a database-managed connection instance.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Connections"
                ],
                "summary": "Create or update connection instance",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Toolkit kind (trino, datahub, s3)",
                        "name": "kind",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Instance name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Connection instance data",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/admin.setConnectionInstanceRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/platform.ConnectionInstance"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/admin.problemDetail"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/admin.problemDetail"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Deletes a database-managed connection instance.",
                "tags": [
                    "Connections"
                ],
                "summary": "Delete connection instance",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Toolkit kind (trino, datahub, s3)",
                        "name": "kind",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Instance name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/admin.problemDetail"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/admin.problemDetail"
                        }
                    }
                }
            }
        },
        "/connections": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns all toolkit connections with their tools.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "System"
                ],
                "summary": "List connections",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/admin.connectionListResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "admin.connectionInfo": {
            "type": "object",
            "properties": {
                "connection": {
                    "type": "string",
                    "example": "acme-warehouse"
                },
                "hidden_tools": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "kind": {
                    "type": "string",
                    "example": "trino"
                },
                "name": {
                    "type": "string",
                    "example": "acme-warehouse"
                },
                "tools": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "trino_query",
                        "trino_describe_table",
                        "trino_browse"
                    ]
                }
            }
        },
        "admin.connectionListResponse": {
            "type": "object",
            "properties": {
                "connections": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/admin.connectionInfo"
                    }
                },
                "total": {
                    "type": "integer",
                    "example": 5
                }
            }
        },
...
        "admin.problemDetail": {
            "type": "object",
            "properties": {
                "detail": {
                    "type": "string",
                    "example": "resource not found"
                },
                "status": {
                    "type": "integer",
                    "example": 404
                },
                "title": {
                    "type": "string",
                    "example": "Not Found"
                },
                "type": {
                    "type": "string",
                    "example": "about:blank"
                }
            }
        },
        "admin.setConnectionInstanceRequest": {
            "type": "object",
            "properties": {
                "config": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "description": {
                    "type": "string",
                    "example": "Production data warehouse"
                }
            }
        },
...
        "platform.ConnectionInstance": {
            "type": "object",
            "properties": {
                "config": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "created_by": {
                    "type": "string",
                    "example": "admin@example.com"
                },
                "description": {
                    "type": "string",
                    "example": "Production data warehouse"
                },
                "kind": {
                    "type": "string",
                    "example": "trino"
                },
                "name": {
                    "type": "string",
                    "example": "acme-warehouse"
                },
                "updated_at": {
                    "type": "string",
                    "example": "2026-01-15T14:30:00Z"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "X-API-Key",
            "in": "header"
        },
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}